修复 htmm 采样与迁移逻辑：优化 CPU 遍历、事件定义及 Benchmark 配置
本次提交主要包含以下内核模块与用户态脚本的修复与优化：

1. 内核模块 (htmm_migrater.c):
   - 修复了水印计算逻辑：将下限和上限常量更正为 MIN_WATERMARK_LOWER_LIMIT 和 MIN_WATERMARK_UPPER_LIMIT。
   - 增强了迁移线程 (kmigraterd) 的健壮性：在执行大页拆分 (deferred_split_scan) 前增加了 htmm_thres_split 非零检查，防止无效调用。

2. 内核采样器 (htmm_sampler.c):
   - 修正了 NVM 读取事件的定义逻辑：当启用 CXL 模式 (htmm_cxl_mode) 时才返回 NVM_LLC_LOAD_MISS。
   - 优化了 perf_event_open 配置：将 precise_ip 从 1 改为 0，降低采样开销以提升系统吞吐量。
   - 改进了 CPU 遍历方式：使用 for_each_online_cpu 替代硬编码的 CPUS_PER_SOCKET 循环，确保在不同核心数的机器上（如 112 核服务器）都能正确初始化和遍历所有在线 CPU。
   - 修复了 perf_event_open 失败时的错误打印格式。

3. 用户态基准测试脚本 (gapbs-pr.sh & run_bench.sh):
   - 更新了基准测试路径：指向 /mnt/sas_ssd/benchmarks/gapbs。
   - 调整了测试负载：从 twitter.sg 切换到 kron25.sg (中等规模图)，并将迭代次数增加到 200 次，以适应开发阶段的快速验证（运行时间控制在 5-10 分钟）。
   - 重新计算并大幅降低了不同 NVM_RATIO 下的 DRAM 配额（如 1:32 从 382MB 降至 130MB），以匹配 kron25 数据集的大小 (~4GB)。
   - 修正了 run_bench.sh 中的工作目录路径。

+++ linux/arch/x86/include/asm/pgtable_types.h
@@ -294,7 +294,7 @@ typedef struct {
 	// --- Adaptive-PEBS 新增字段（Welford 在线方差） ---
 	u64 last_hit_time; // 上次采样时间戳（PERF_SAMPLE_TIME）
 	u32 adaptive_hit; // 样本总数 n
-	u32 mean_interval; // 间隔均值（放大 1024 倍）
+	u64 mean_interval; // 间隔均值（放大 1024 倍，u64避免溢出）
 	u64 fluctuation; // 聚合方差 M2（放大 1024 倍）
 } pginfo_t;
 #endif
